+++ backend/src/main/java/com/example/coworking/repository/EspacioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.example.coworking.model.Espacio;
 
-public interface EspacioRepository extends JpaRepository<Espacio, Long> {
+public interface EspacioRepository extends JpaRepository<Espacio, Integer> {
 
     @Query("SELECT DISTINCT e.tipo FROM Espacio e")
     List<String> findDistinctTipos();
