+++ backend/src/main/java/com/example/coworking/repository/EspacioServicioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.coworking.model.EspacioServicio;
 
-public interface EspacioServicioRepository extends JpaRepository<EspacioServicio, Long> {
+public interface EspacioServicioRepository extends JpaRepository<EspacioServicio, Integer> {
 
 }
