+++ backend/src/main/java/com/example/coworking/repository/PagoRepository.java
@@ -2,6 +2,6 @@ package com.example.coworking.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.coworking.model.Pago;
 
-public interface PagoRepository extends JpaRepository<Pago, Long> {
+public interface PagoRepository extends JpaRepository<Pago, Integer> {
 
 }
