+++ backend/src/main/java/com/example/coworking/repository/TipoEspacioRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.coworking.model.TipoEspacio;
 
-public interface TipoEspacioRepository extends JpaRepository<TipoEspacio, Long> {
+public interface TipoEspacioRepository extends JpaRepository<TipoEspacio, Integer> {
 }
